# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.01"
  title: App
  description: API Documentation
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
#host: Taked automatically from site
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/private/movements:
    x-swagger-router-controller: movement.controller
    get:
        description: Returns the list of all movements
        operationId: movements
        tags: [movements]
        security:
          - Bearer: []
        consumes:
          - application/json
        produces:
          - application/json
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/movementsResponse"
            # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
...
  /api/private/discounts/next:
    x-swagger-router-controller: discount.controller
    get:
        description: Returns the list of all next discounts
        operationId: proximos
        tags: [discounts]
        security:
          - Bearer: []
        consumes:
          - application/json
        produces:
          - application/json
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/discountsResponse"
            # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"          
...
  /api/private/users:
    x-swagger-router-controller: user.controller
    get:
      security:
        - Bearer: []
      description: Get all users
      operationId: getAll
      tags: [Users]
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/usersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
    put:
      security:
        - Bearer: []
      description: Update user roles
      operationId: putUserRoles
      tags: [Users]
      consumes:
        - application/json
      produces:
        - application/json
      parameters: 
      - name: user
        in: body
        description: Fields to update the user roles
        required: true
        schema:
          $ref: '#/definitions/userRequest'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/basicResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
...
definitions:
  movementsResponse:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/movementResponse'
  movementResponse:
    type: object
    properties:
      date: 
        type: object
      ammount:
        type: number
      categoria:
        type: string
      descripcion:
        type: string  
  discountsResponse:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/discountResponse'
  discountResponse:
    type: object
    properties:
      categoria:
        type: string
      vencimiento:
        type: object
        example: "One of Vigente, Vencido, Canjeado" 
      percentage: 
        type: string
        example: "5%, 10%, 15%"
      estado: 
        type: string
...
  userRequest:
    type: object
    properties:
      dni:
        type: string
        example: User dni
      displayName:
        type: string
        example: User fullname
      password:
        type: string
        example: User password
      roles: 
        type: array
        example: User roles
        items: 
          type: string
...
  errorResponse:
    type: object
    properties:
      message:
        type: string
  basicResponse:
    type: object
    properties:
      message:
        type: string
  usersResponse:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/userResponse'
  userResponse:
    type: object
    required:
      - _id
      - email
      - displayName
      - roles
    properties:
      _id:
        type: string
        example: User id
      email:
        type: string
        example: User email
      displayName:
        type: string
        example: User fullname
      roles: 
        type: array
        example: User roles
        items: 
          type: string